Add no-op metric implementations and options
Sketch out the overall shape of the package: this is enough to get
everything compiling, and limits the scope of subsequent commits.

First, add no-op implementations of the Counter, Gauge, Histogram, and
the corresponding vector types. Since these are structs rather than
interfaces, it's simple to make nil pointers no-ops. The actual metric
behavior is omitted, and will be added in future commits.

Then, add sketches of the Registry and Controller types. The Registry is
the central metrics object, similar to Tally's Scope. Like the
individual metrics, nil Registries are usable no-op implementations. The
actual Registry and Controller implementations will be added in future
commits.

Finally, add the user-facing options for metric construction. These are
part of the public Registry API. Note that we specifically avoid
functional options here to avoid unnecessary allocations.

Also, keep in mind that this commit only unit tests the no-op code
paths. The actual metrics code will be tested in the commits that
replace the temporary logic included here.