Enforce metric uniqueness within a registry
Within a single registry, we want metrics to be unique - it greatly
simplifies the Prometheus exposition code, and it makes it dead simple
to identify the source of metrics during an incident.

We do this by introducing a `metadata` struct, which is nothing more
than a validated and further-processed representation of the data
contained in the user-facing options structs. Using the metadata, the
registry can enforce a sane set of uniqueness constraints (explained in
a comment on the new core registry type, and explored in detail in an
extensively-commented series of unit tests).

This commit also introduces the `metric` interface, which Counters,
Gauges, Histograms, and their vector forms all satisfy. To start with,
this interface only requires that metrics be able to provide their own
metadata. Future commits will require that metrics also be capable of
serializing themselves to protobuf and pushing to Tally.