Add basic testing helpers
As a prerequisite for a full-featured `metricstest` package and as a
convenience for testing this package, this commit adds the ability for
controllers to produce a snapshot of all the metrics in their paired
registry. For now, all snapshots are empty; as implementations of the
various metric types are added, we'll also add the required snapshotting
logic.

Note that we're adding the snapshot-producing method to the controller,
not the registry itself: since the snapshot is global, it belongs on the
controller. This will become clearer in a subsequent commit, when we add
the ability to create sub-registries with a consistent set of labels.

This code is exercised in the unit tests for the various metric types,
which will be added in subsequent commits.