Add interfaces for StatsD and M3 integration
Add the `push` sub-package, which contains nothing but interface
definitions for integration with push-based telemetry systems (like
StatsD or M3) and a no-op implementation of the interface.

These interfaces will be referenced from the main package, so we want
them to live in a separate package from any references to Tally. This
lets programs that use `net/metrics` but not Tally avoid pulling in any
of Tally or its dependencies.