Add a Tally-backed implementation of push.Target
Add a `tallypush` package with a Tally-based implementation of the
`push.Target` interface. Keeping the concrete implementation in a
separate package from the interfaces ensures that only applications that
actually push to Tally have to compile the Tally code.

This commit includes simple unit tests for the Tally pusher types. A
more thorough integration test will be included with the commit that
integrates these push subpackages with the main `metrics.Controller`
struct.