Add push support to all metric types
Building on the `push` and `tallypush` subpackages, add push support to
all metrics. This requires expanding the `metric` interface to include a
method that each metric can use to push itself, and then wiring these
methods together from the controller.

Note that none of the push methods need to be safe for concurrent use,
because the controller guarantees that only one push ticker runs at a
time.