Add support for Prometheus's protobuf exposition
This commit makes the controller an HTTP handler, which uses portions of
the official Prometheus client to expose all the registry's metrics for
scraping or debugging. The Prometheus client's HTTP handler is pretty
nice - it automatically does some content-type negotiation, so it dumps
plain text by default but also supports negotiating up to the binary
protobuf encoding.

This requires one final extension to the internal `metric` interface:
all metrics must now be able to serialize themselves to Prometheus's
`MetricFamily` protobuf type. This is an internal detail, so we can
definitely change if the Prometheus exposition format changes (which it
likely will - there's a group of folks working on a common exposition
format for a whole variety of metrics systems).

This is tested via an end-to-end integration test. The fixture for the
integration test was generated by the official Prometheus client
library; in a post-1.0 commit, I'll add the code that generated it and
verify that it also matches the fixture exactly.