Rename Opts to Spec
Before cutting a 1.0, we should carve out some space for future
registry- and controller-level options. Unfortunately, there isn't a
great name for this type available, since we've used `Opts` for metric
metadata.

This commit renames the `Opts` and `HistogramOpts` types to `Spec` and
`HistogramSpec`, which leaves `Option` free for the more typical
functional option pattern.