Unify Registry and Controller
Returning two top-level structs from `New` is unusual, and it
makes it harder to choose good names for these types. Instead, this
commit reorganizes and renames so that `New` returns a single struct.

First, we rename `Controller` to `Root` (since it contains top-level
functionality like snapshotting, managing the push loop, and exposing
HTTP endpoints), and rename `Registry` to `Scope` (which aligns us more
closely with Tally's terminology). Then, we expose the scope via a
method on the root and change `New` to return only one struct.

For cleanliness, we also rename the `RegistrySnapshot` to `RootSnapshot`
and rename local variables.