Add type hint to histograms for superior tally integration
Summary: Currently, all histograms created through netmetrics will
always be recorded as values.  This mostly affects the bucket name tags
that are recorded with the histograms.  Currently, the histograms are
being recorded as "1.00000-2.00000" where ideally they would be
"1ms-2ms".

This PR keeps the "Value" as the default for histograms, but adds a Type
field to the HistogramSpec that we thread through to the
push.HistogramSpec so that our tally integration can do the right thing
based on the type of histogram.

Biggest part of this PR that's iffy for me is the repeated const values
in the metrics and push packages, open to suggestions to isolate that.