Migrate to Go modules (#29)
This migrates net/metrics to Go modules, significantly simplifying the
build set up in the process. The glide.yaml is left in place to provide
constraints for legacy ecosystems.

The `make bench` target was dropped because this repository does not
contain any benchmarks.

The `staticcheck` linter was added and issues found by it fixed.