Improve histogram push (#38)
Iterating an array to access a map for every element is expensive.
Therefore, use a slice internally. The old API would be slower (~25%),
but the new API is 3.35x faster. This way we are able to support any
library that could be using it, but internally we use the new API to get
the benefits from it.